requirements.txt

fastapi
numpy
pandas
pydantic
requests
snowflake-connector-python
streamlit
uvicorn

